feat: loading data in service for better performance
