+++ packages/vendure-plugin-metrics/src/api/schema.graphql.ts
@@ -1,5 +1,10 @@
 import gql from 'graphql-tag';
 
+// This is only used by codegen so it knows DateTime is a custom scalar
+const scalars = gql`
+  scalar DateTime
+`;
+
 export const schema = gql`
   type AdvancedMetricSummary {
     code: String!
