+++ packages/vendure-plugin-metrics/src/index.ts
@@ -1,7 +1,7 @@
 export * from './metrics.plugin';
-export * from './api/metrics.service';
+export * from './services/metrics.service';
 export * from './api/metrics.resolver';
 export * from './ui/generated/graphql';
-export * from './api/metrics/average-order-value';
-export * from './api/metrics/revenue-per-product';
-export * from './api/metrics/units-sold-metric';
+export * from './metrics/average-order-value';
+export * from './metrics/revenue-per-product';
+export * from './metrics/units-sold-metric';
