+++ packages/vendure-plugin-metrics/src/metrics.plugin.ts
@@ -3,18 +3,18 @@ import { PluginCommonModule, VendurePlugin } from '@vendure/core';
 import path from 'path';
 import { schema } from './api/schema.graphql';
 import { MetricsResolver } from './api/metrics.resolver';
-import { MetricsService } from './api/metrics.service';
-import { MetricStrategy } from './api/metric-strategy';
+import { MetricsService } from './services/metrics.service';
+import { MetricStrategy } from './services/metric-strategy';
 import { PLUGIN_INIT_OPTIONS } from './constants';
-import { RevenuePerProduct } from './api/metrics/revenue-per-product';
-import { AverageOrderValueMetric } from './api/metrics/average-order-value';
-import { UnitsSoldMetric } from './api/metrics/units-sold-metric';
+import { RevenuePerProduct } from './metrics/revenue-per-product';
+import { AverageOrderValueMetric } from './metrics/average-order-value';
+import { UnitsSoldMetric } from './metrics/units-sold-metric';
 
 export interface MetricsPluginOptions {
   /**
    * The enabled metrics shown in the widget.
    */
-  metrics: MetricStrategy<any>[];
+  metrics: MetricStrategy[];
   /**
    * The number of past months to display in the metrics widget.
    * If your shop has a lot of orders, consider using only the last 3 months for example.
