+++ packages/vendure-plugin-metrics/src/metrics/revenue-per-product.ts
@@ -1,20 +1,21 @@
 import {
   Injector,
   Logger,
+  Order,
   OrderLine,
   ProductVariant,
   RequestContext,
   TransactionalConnection,
 } from '@vendure/core';
-import { AdvancedMetricType } from '../../ui/generated/graphql';
-import { MetricStrategy, NamedDatapoint } from '../metric-strategy';
-import { loggerCtx } from '../../constants';
+import { AdvancedMetricType } from '../ui/generated/graphql';
+import { MetricStrategy, NamedDatapoint } from '../services/metric-strategy';
+import { loggerCtx } from '../constants';
 
 /**
  * Calculates the revenue generated by each product variant.
  * calculates the total revenue if no variants are provided.
  */
-export class RevenuePerProduct implements MetricStrategy<OrderLine> {
+export class RevenuePerProduct implements MetricStrategy {
   readonly metricType: AdvancedMetricType = AdvancedMetricType.Currency;
   readonly code = 'revenue-per-product';
   readonly allowProductSelection = true;
@@ -27,60 +28,18 @@ export class RevenuePerProduct implements MetricStrategy<OrderLine> {
     return entity.order.orderPlacedAt ?? entity.order.updatedAt;
   }
 
-  async loadEntities(
-    ctx: RequestContext,
-    injector: Injector,
-    from: Date,
-    to: Date,
-    variants: ProductVariant[]
-  ): Promise<OrderLine[]> {
-    let skip = 0;
-    const take = 1000;
-    let hasMoreOrderLines = true;
-    const lines: OrderLine[] = [];
-    while (hasMoreOrderLines) {
-      let query = injector
-        .get(TransactionalConnection)
-        .getRepository(ctx, OrderLine)
-        .createQueryBuilder('orderLine')
-        .leftJoinAndSelect('orderLine.order', 'order')
-        .leftJoin('order.channels', 'channel')
-        .where(`channel.id=:channelId`, { channelId: ctx.channelId })
-        .andWhere('order.orderPlacedAt BETWEEN :fromDate AND :toDate', {
-          fromDate: from.toISOString(),
-          toDate: to.toISOString(),
-        })
-        .offset(skip)
-        .limit(take);
-      if (variants.length) {
-        query = query
-          .leftJoin('orderLine.productVariant', 'productVariant')
-          .addSelect(['productVariant.sku', 'productVariant.id'])
-          .andWhere(`productVariant.id IN(:...variantIds)`, {
-            variantIds: variants.map((v) => v.id),
-          });
-      }
-      const [items, totalItems] = await query.getManyAndCount();
-      lines.push(...items);
-      Logger.debug(
-        `Fetched order lines ${skip}-${skip + take} for metric ${
-          this.code
-        } for channel${ctx.channel.token}`,
-        loggerCtx
-      );
-      skip += items.length;
-      if (lines.length >= totalItems) {
-        hasMoreOrderLines = false;
-      }
-    }
-    return lines;
-  }
-
   calculateDataPoints(
     ctx: RequestContext,
-    lines: OrderLine[],
+    orders: Order[],
     variants: ProductVariant[]
   ): NamedDatapoint[] {
+    // Map orders to lines, but st the order object on each line
+    const lines = orders
+      .map((order) => {
+        order.lines.forEach((l) => (l.order = order));
+        return order.lines;
+      })
+      .flat();
     if (!variants.length) {
       // Return total revenue
       let legendLabel = 'Total Revenue';
